third_party/WebKit/LayoutTests/platform/mac-retina/virtual/outofblink-cors/external/wpt/service-workers/service-worker/activation.https-expected.txt

This is a testharness.js-based test.
FAIL loss of controllees triggers activation assert_equals: expected object "[object ServiceWorker]" but got object "[object ServiceWorker]"
PASS finishing a request triggers activation
PASS skipWaiting bypasses no controllee requirement
PASS finishing a request triggers unregister
Harness: the test ran to completion.
